pyproject.toml
[project]
name = "dc-emv-cot"
version = "0.1.0"
description = "Dual-compilation executable metacognitive verification experiments"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.2.0",
    "transformers>=4.46.0",
    "datasets>=2.18.0",
    "accelerate>=0.29.0",
    "hydra-core>=1.3.2",
    "wandb>=0.17.0",
    "optuna>=3.6.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "tqdm>=4.66.0",
    "sentencepiece>=0.2.0",
]
